clazz renamed to classes
