+++ respect-app-compose/src/commonMain/kotlin/world/respect/app/view/classes/ClassesScreen.kt
@@ -1,11 +1,11 @@
-package world.respect.app.view.clazz
+package world.respect.app.view.classes
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import world.respect.shared.viewmodel.clazz.ClazzViewModel
 
 @Composable
-fun ClazzScreen(
+fun ClassesScreen(
     navController: NavHostController,
     viewModel: ClazzViewModel
 ) {
