+++ respect-lib-shared/src/commonMain/composeResources/values/strings.xml
@@ -14,7 +14,7 @@
     <string name="apps">Apps</string>
     <string name="app">App</string>
     <string name="assignments">Assignments</string>
-    <string name="clazz">Classes</string>
+    <string name="classes">Classes</string>
     <string name="report">Reports</string>
     <string name="select_app">Select app</string>
     <string name="enter_link">Enter link</string>
